backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
asyncpg
databases
folium
qrcode
pillow